changes the structure
