--- src/components/PeerJSGroupRoom/peerJSGroupRoom.css
@@ -1,9 +1,0 @@
-.video-grid{
-    width: 300px;
-    height: 300px;
-}
-
-.video {
-    width: 100%;
-    height: 100%;
-}
