+++ src/pages/PeerJSGroupRoom/peerJSGroupRoom.css
@@ -1,0 +1,12 @@
+.video {
+    width: 100%;
+    height: 100%;
+}
+.remote-user-video-grid{
+    width: 250px;
+    height: 250px;
+}
+.current-user-video-grid{
+    width: 400px;
+    height: 400px;
+}
